// src/app/home/models/Proprietario.ts
export class Proprietario {
    proprietario: string;
    cpf: string;
    rg: string;
    tipo1: string;
    logradouro1: string;
    numero1: string;
    bairro1: string;
    cep1: string;
    complemento1: string;

    setor1: string;
    inscricao: string;
    tipo2: string;
    logradouro2: string;
    numero2: string;
    bairro2: string;
    cep2: string;
    complemento2: string;

    frente: string;
    fundo: string;
    ladodireito: string;
    ladoesquerdo: string;
    areaunidconstr: string;
    areatotcostedif: string;
}